requirements.txt
python-chess==0.23.6
requests==2.19.1